requirements.txt
Django >4.2,<4.3
python-decouple
pillow
djangorestframework
markdown       
django-filter
drf_spectacular
djangorestframework-simplejwt

# email template module
django-mail-templated

# testing module
pytest
pytest-django
black
flake8